Added Euler Average and fixed setting function. Added an average euler function which averages the eulers, which i actually think is a great fit for rotation averaging. If eulers had some strength that i dont know about, this might be it.
The only praise i ever hear for euler angles is that they're easy to
understand and manipulate but people go on and on about "order" this and
"gimbal lock" that. With the way that we're averaging the euler angles,
i can't see how big of an issue gimbal lock could possibly be, and bevy
handles all order issues that everyone complains about. When it comes to
averaging, i look at the different rotation formats; Quaternions, scaled
axes, euler angles etc. the same way you would look at lerping different
colour formats, just gives a different result, and the way that euler
angles just are, i think that they're actually good for this specific
use case. i'll have to try out scaled axes now that the issues are fixed
and see what i think.

also fixed the set_targets function, i don't know what i was thinking,
but i and the sort of direction i was taking the function was confused.
It was only querying on added transforms, while running every frame.