Added Covariance Setting, function + minor rework Firstly reworked the way that functions are selected, now a setting on each out instead of a global resource, allowing me to compare the differences between each method.
Added this eigen covariance method which i found and pieced together
from papers and posts online it is probably the best of the now current
3 methods, however it has to converge which is annoying.

After comparing each one, i realised that the euler method definately
does suffer from gimbal lock, although sometimes it works as an
advantage, sometimes not. when there are two targets, and one has
rotation only on the x, while the other has rotation around only the y.
the other two methods tend to swing outwards, which looks pretty weird
if you ask me. The Euler method won't do this though, and i think it's
because of the gimbal lock. however there are other situations where
this effect works against it. The euler method tends to jump around more
when the range of the targets gets larger, more so then the other two.

However, all methods do this flip thing, so perhaps i should do a copy
of the Covariance and make it based on the upward direction instead,
with a 3x3 covariance matrix instead. that might work better. also when
i port this to the Ik plugin, I'll probably use the rotation / direction
of the parent object in question as the base eigenvector. maybe that way
it'll be less likely to flip about, that's my theory anyway.